Update: modularization
