+++ channels/grid.py
@@ -1,113 +1,7 @@
-from collections import deque
-from copy import deepcopy
-
-def BFS_for_nearest_unknown_cell(low_res_grid, new_width, new_height, start_x, start_y):
-    directions = [(0, 1, 'right'), (0, -1, 'left'), (1, 0, 'down'), (-1, 0, 'up')]
-    start_idx = start_y * new_width + start_x
-    queue = deque([(start_x, start_y)])
-    visited = set()
-    visited.add(start_idx)
-    while queue:
-        x, y = queue.popleft()
-        for dx, dy, action in directions:
-            nx, ny = x + dx, y + dy
-            nidx = ny * new_width + nx
-            if 0 <= nx < new_width and 0 <= ny < new_height:
-                if low_res_grid[nidx] == -1: #found unknown
-                    return (nx, ny)
-                if low_res_grid[nidx] == 0 and nidx not in visited:
-                    queue.append((nx, ny))
-                    visited.add(nidx)
-    return None
-
-from hyperon.ext import register_atoms
-from hyperon import *
-import time
-
-NAV_STATE_FAIL = "FAIL"
-NAV_STATE_BUSY = "BUSY"
-NAV_STATE_SUCCESS = "SUCCESS"
-NAV_STATE = NAV_STATE_SUCCESS #"busy", "success" and "fail" (initially set to success)
-def NAV_STATE_SET(NAV_STATE_ARG):
-    global NAV_STATE
-    NAV_STATE = NAV_STATE_ARG
-    return None
-
-class PatternOperation(OperationObject):
-    def __init__(self, name, op, unwrap=False, rec=False):
-        super().__init__(name, op, unwrap)
-        self.rec = rec
-    def execute(self, *args, res_typ=AtomType.UNDEFINED):
-        return super().execute(*args, res_typ=res_typ)
-
-def wrapnpop(func):
-    def wrapper(*args):
-        a = [str("'"+arg) if arg is SymbolAtom else str(arg) for arg in args]
-        res = func(*a)
-        return [res]
-    return wrapper
-
-MeTTaROS2Command = ""
-def call_bridgeinput(*a):
-    global runner, MeTTaROS2Command
-    tokenizer = runner.tokenizer()
-    cmd = str(a[0])
-    parser = SExprParser(f"(MeTTaROS2Command {cmd})")
-    MeTTaROS2Command = cmd
-    return parser.parse(tokenizer)
-
-def space_init():
-    global runner
-    with open("space.metta", "r") as f:
-        metta_code = f.read()
-    runner = MeTTa()
-    call_bridgeinput_atom = G(PatternOperation('bridgeinput', wrapnpop(call_bridgeinput), unwrap=False))
-    runner.register_atom("bridgeinput", call_bridgeinput_atom)
-    runner.run(metta_code)
-
-currentTime = 0
-start_time = time.time()
-
-def space_tick(node):
-    global currentTime, MeTTaROS2Command
-    cmd = MeTTaROS2Command
-    MeTTaROS2Command = ""
-    if cmd == "right":
-        node.start_navigation_by_moves("right")
-    if cmd == "left":
-        node.start_navigation_by_moves("left")
-    if cmd == "up":
-        node.start_navigation_by_moves("up")
-    if cmd == "down":
-        node.start_navigation_by_moves("down")
-    if cmd.startswith("(goto (coordinates "):
-        x_y = cmd.split("(goto (coordinates ")[1].split(")")[0]
-        x = int(x_y.split(" ")[0])
-        y = int(x_y.split(" ")[1])
-        node.start_navigation_to_coordinate((x, y))
-    alldetections = deepcopy(node.previous_detections)
-    objects = "("
-    if "self" in node.previous_detections:
-        (t, object_grid_x, object_grid_y, origin_x, origin_y) = node.previous_detections["self"]
-        x_y_unknown = BFS_for_nearest_unknown_cell(node.low_res_grid, node.new_width, node.new_height, object_grid_x, object_grid_y)
-        if x_y_unknown:
-            (x_unknown,y_unknown) =  x_y_unknown
-            alldetections["unknown"] = (time.time(), x_unknown, y_unknown, origin_x, origin_y)
-    print(alldetections)
-    for category in alldetections:
-        (t, object_grid_x, object_grid_y, _, __) = alldetections[category]
-        SEXP = f"(detection {category} (coordinates {object_grid_x} {object_grid_y}))"
-        objects += SEXP
-        if category == "self":
-            SELF_position = (object_grid_x, object_grid_y)
-    objects += ")"
-    currentTime += 1
-    elapsedTime = round(time.time() - start_time, 2)
-    print("NAV_STATE", NAV_STATE, runner.run(f"!(Step {currentTime} {elapsedTime} {NAV_STATE} {objects})"))
-
-space_init()
+from mettabridge import *
 
 #!/usr/bin/env python3
+import time
 import rclpy
 import numpy as np
 import math
